(RE-5292) Depend on pe-client-tools for colored gem
Prior to this commit, pe-client-tools and pe-r10k
are unable to be co-installed on the master, since
both install the same colored gem to the same path.

This commit removes the colored gem from pe-r10k
and updates pe-r10k's vanagon project configuration
so that it will depend on pe-client-tools, which
will provide the required colored gem.