(maint) Check that r10k works during build
In order to avoid breakages, this commit adds a simple check to the r10k
install step to ensure that r10k at least loads. It would be preferable
to have the check in a check step, but due to
https://tickets.puppetlabs.com/browse/VANAGON-58 the check happens at
the wrong time of the build for gems. If that ticket ever gets resolved
the check can move.