(maint) Update r10k component config to enable building from SHA
Previously, the r10k config assumed that the component would only ever
refer to a tagged version of the gem. The r10k Jenkins pipelines,
however, are vanagon standard and designed to run post-merge to the
component on unreleased SHAs. This commit updates the config to remove
direct references to the gem's version, which mirrors how we handle
other gems we actively develop, in other vanagon projects (see the
Puppet Resource API in Puppet Agent).