(maint) roll back rugged/libgit2 versions
Previously we had to move forward from our no-longer maintained versions
of libgit2 & rugged so we could begin building packages on RHEL8 (an
issue with ruby 2.5 and more recent llvms caused rugged to fail to
compile pre-0.27.0.

In rolling forward we moved to the latest release of both libgit2/rugged
(0.28.1). However the 0.28 series has only been out for two months at
this point and still may be working some kinks out. We specifically are
seeing issues with their new re-written handing of proxies.

We may need to review how we configure proxies in r10k to account for
libgit2 0.28.0+ proxy handling, and that effort is unknown, at a minimum
it in its current configuration it no longer gives any proxy related
errors or warning on misconfiguration. That classifies at least as a
regression in usability. Unfortunately, with r10k's current low priority
rolling to a well supported, stable, and known good build is the preferred
fix for now.